refactor(native-ui): rename StepListItemWrapper to just StepListItem
